UserService/requirements.txt
aio-pika==9.5.7
aiormq==6.9.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
multidict==6.6.3
pamqp==3.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
redis==6.4.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1